requirements.txt
fastapi
uvicorn[standard]
PyMuPDF
sentence-transformers
chromadb
requests
python-dotenv